+++ source/include (local)/sensorID.h
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include "typedefs.h"
 
-int IoTLib_sensorid_get_key_index(const IoTLib_SensorID* keyArray, const int arrayLength, const IoTLib_SensorID keyToSearchFor);
+size_t IoTLib_sensorid_get_key_index(const IoTLib_SensorID* keyArray, const size_t arrayLength, const IoTLib_SensorID keyToSearchFor);
